refactor: move `RecordCard.vue` back to views/ & add `Recent.vue` back
